// src/types/reports/reports.ts
export interface IReportsStore {
  label: string;
  value: string;
}

export const DEFAULT_START_DATE = new Date("2024-01-01");
export const DEFAULT_END_DATE = new Date("2024-06-30");
// ...
export interface IOrderItem {
  productName: string;
  quantity: number;
  unitCost: number;
  totalCost: number;
  ourCost: number;
  profit: number;
  isOrderHeader?: boolean;
  isTotal?: boolean;
  isOrder?: boolean;
}

export interface IOrderTotals {
  quantity: number;
  unitCost: number;
  totalCost: number;
  ourCost: number;
  profit: number;
}

export interface IOrder {
  orderId: string;
  orderTotal: number;
  orderCost: number;
  shippingCost: number;
  shipping: number;
  profit: number;
  items: IOrderItem[];
  totals: IOrderTotals;
}

export interface IOrderTableRow extends IOrderItem {
  isOrderHeader?: boolean;
  isTotal?: boolean;
  orderId?: string;
  orderTotal?: number;
  orderCost?: number;
  shippingCost?: number;
  shipping?: number;
}